[Feat] DataAccessProvider 설정 (#29) - 구조가 맘에 안듦
