+++ OverTheRainbow/OverTheRainbow/DataAccess/Realm/RealmConfig.swift
@@ -9,15 +9,18 @@
 import Foundation
 import RealmSwift
 
-class RealmProvider {
-
+class RealmConfig: DataAccessConfig {
+    public static let config = RealmConfig()
+    
     private let realm: Realm = try! Realm()
     
     private func getRepository() -> RealmRepository {
         return RealmRepository(realm: self.realm)
     }
     
-    public func getService() -> RealmService {
+    public func getService() -> DataAccessService {
         return RealmService(realm, self.getRepository())
     }
+    
+    private init() {}
 }
